NFC: Move HoistableOpInterface external registration to Util
The op hoisting interface doesn't require any special casting
logic when making hoisting decisions, so we can register them like
all other Util interfaces. We still need to put up with the
spooky-action-at-a-distance registration for tensor types because:

a) Sub-byte dense constants aren't representable in MLIR
b) There is no tensor.reinterpret_cast op to replace flow.bitcast
