+++ compiler/src/iree/compiler/Dialect/Util/Transforms/test/BUILD.bazel
@@ -25,6 +25,7 @@ iree_lit_test_suite(
             "fold_globals.mlir",
             "fuse_globals.mlir",
             "hoist_into_globals.mlir",
+            "hoist_into_globals_linalg.mlir",
             "import_resources.mlir",
             "ipo.mlir",
             "outline_constants.mlir",
