+++ compiler/src/iree/compiler/Dialect/Util/Transforms/test/CMakeLists.txt
@@ -23,6 +23,7 @@ iree_lit_test_suite(
     "fold_globals.mlir"
     "fuse_globals.mlir"
     "hoist_into_globals.mlir"
+    "hoist_into_globals_linalg.mlir"
     "import_resources.mlir"
     "ipo.mlir"
     "outline_constants.mlir"
