+++ compiler/src/iree/compiler/GlobalOptimization/test/CMakeLists.txt
@@ -24,7 +24,6 @@ iree_lit_test_suite(
     "fuse_silu_horizontal_matmul.mlir"
     "generalize_named_ops.mlir"
     "global_loop_invariant_code_motion.mlir"
-    "hoist_into_globals_linalg.mlir"
     "infer_numeric_narrowing.mlir"
     "materialize_homogeneous_encodings.mlir"
     "optimize_numerics.mlir"
